+++ V5.0/ZDK/src/nl/zeesoft/zdk/midi/LfoTarget.java
@@ -3,14 +3,14 @@ package nl.zeesoft.zdk.midi;
 public class LfoTarget {
 	public int		channel		= 0;
 	public String	property		= Inst.FILTER;
-	public int		percentage		= 20;
+	public float	percentage		= 20;
 	public boolean	invert			= false;
 	
 	public LfoTarget() {
 		
 	}
 	
-	public LfoTarget(int channel, String property, int percentage, boolean invert) {
+	public LfoTarget(int channel, String property, float percentage, boolean invert) {
 		this.channel = channel;
 		this.property = property;
 		this.percentage = percentage;
